requirements.txt
python-dotenv
openai==1.51.0
langchain==0.3.4
langchain-openai==0.2.2
langchain-anthropic==0.2.3
langchain-chroma==0.1.4
pillow
langserve==0.3.0
langsmith==0.1.131
streamlit==1.39.0
chainlit==1.2.0
gradio==4.44.1
langchain-milvus==0.1.5
pypdf==5.0.1
boto3>=1.28.57
awscli>=1.29.57
botocore>=1.31.57